+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The **patch** part changes is incremented if multiple releases happen the same m
 
 ### Added
 
+* autosysadmin-agent: upstream release 24.02.3
 * certbot: Renewal hook for NRPE
 * kvm-host: add minifirewall rules if DRBD interface is configured
 
@@ -26,6 +27,8 @@ The **patch** part changes is incremented if multiple releases happen the same m
 ### Fixed
 
 ### Removed
+
+* autosysadmin: replaced by autosysadmin-agent
 
 ### Security